# This should improve string performance in ondemand by making the string processing runtime dispatched.