--- a/include/simdjson/ppc64/implementation.h
+++ b/include/simdjson/ppc64/implementation.h
@@ -26,6 +26,7 @@ public:
                                          size_t &dst_len) const noexcept final;
   simdjson_warn_unused bool validate_utf8(const char *buf,
                                           size_t len) const noexcept final;
+  simdjson_warn_unused uint8_t *parse_string(const uint8_t *src, uint8_t *dst) const noexcept final;
 };
 
 } // namespace ppc64