… include/simdjson/generic/stringparsing.h