@lemire 2022-06-22
I think I did it wrong. Having it as a separate dispatched function probably carries too much overhead. Let us investigate.